tophat_scripts_python/requirements.txt
playwright==1.44.0
PyYAML==6.0.1

